Refactored resident and address, altered tests, and modified server
Taking data from Rick Leckinger <richard.leckinger@greens.org.nz>, I found his data had extra fields not currently used and were missing ones that were being used.  Added the fields into Address and Resident and removed DOB from Resident, as until that data is available, there is no point in having it there.  Could have kept it, but decided not to for reasons previously said.

As a result of this refactoring the tests had to be refactored, as thay were relying on constructors that were no longer valid, and were testing fields that no longer existed.

The server SQL DB stuff had to be altered, to take into account the Resident and Address data change.  There is also a effort being made to populate the database with data programatically.