Reduced the complexity of the server code and added a map to the app
Using router for REST and a static file server to serve the app html and dart code.  This way I don't have to create a route for every single file that is needed, and just leave this to the static file server instead.

Created a map polymer element to show addresses and give a way of selecting which ones to do and which ones have been visited.